+++ tmt/base.py
@@ -59,6 +59,14 @@ import tmt.utils
 import tmt.utils.git
 import tmt.utils.jira
 from tmt.checks import Check
+from tmt.container import (
+    SerializableContainer,
+    SpecBasedContainer,
+    container,
+    container_field,
+    container_fields,
+    field,
+    )
 from tmt.lint import LinterOutcome, LinterReturn
 from tmt.result import Result, ResultInterpret
 from tmt.utils import (
@@ -67,14 +75,9 @@ from tmt.utils import (
     EnvVarValue,
     FmfContext,
     Path,
-    SerializableContainer,
     ShellScript,
-    SpecBasedContainer,
     WorkdirArgumentType,
-    container_field,
-    container_fields,
     dict_to_yaml,
-    field,
     normalize_shell_script,
     verdict,
     )
@@ -141,7 +144,7 @@ class _RawFmfId(TypedDict, total=False):
 
 
 # An internal fmf id representation.
-@dataclasses.dataclass
+@container
 class FmfId(
         SpecBasedContainer[_RawFmfId, _RawFmfId],
         SerializableContainer,
@@ -427,7 +430,7 @@ class _RawDependencyFmfId(_RawFmfId):
     type: Optional[str]
 
 
-@dataclasses.dataclass
+@container
 class DependencyFmfId(
         FmfId,
         # Repeat the SpecBasedContainer, with more fitting in/out spec type.
@@ -524,7 +527,7 @@ class _RawDependencyFile(TypedDict):
     pattern: Optional[list[str]]
 
 
-@dataclasses.dataclass
+@container
 class DependencyFile(
         SpecBasedContainer[_RawDependencyFile, _RawDependencyFile],
         SerializableContainer,
@@ -681,7 +684,7 @@ def _normalize_link(key_address: str, value: _RawLinks, logger: tmt.log.Logger) 
     return Links(data=value)
 
 
-@dataclasses.dataclass(repr=False)
+@container(repr=False)
 class Core(
         tmt.utils.ValidateFmfMixin,
         tmt.utils.LoadFmfKeysMixin,
@@ -1102,7 +1105,7 @@ class Core(
         return self.link.has_link(needle=needle)
 
 
-@dataclasses.dataclass(repr=False)
+@container(repr=False)
 class Test(
         Core,
         tmt.export.Exportable['Test'],
@@ -1786,7 +1789,7 @@ def expand_node_data(data: T, fmf_context: FmfContext) -> T:
     return data
 
 
-@dataclasses.dataclass(repr=False)
+@container(repr=False)
 class Plan(
         Core,
         tmt.export.Exportable['Plan'],
@@ -2798,7 +2801,7 @@ class StoryPriority(enum.Enum):
         return self.value
 
 
-@dataclasses.dataclass(repr=False)
+@container(repr=False)
 class Story(
         Core,
         tmt.export.Exportable['Story'],
@@ -3593,7 +3596,7 @@ class Tree(tmt.utils.Common):
                 logger=logger)
 
 
-@dataclasses.dataclass
+@container
 class RunData(SerializableContainer):
     root: Optional[str]
     plans: Optional[list[str]]
@@ -4424,7 +4427,7 @@ class LinkNeedle:
         return False
 
 
-@dataclasses.dataclass
+@container
 class Link(SpecBasedContainer[Any, _RawLinkRelation]):
     """
     An internal "link" as defined by tmt specification.
