+++ tmt/checks/__init__.py
@@ -1,4 +1,3 @@
-import dataclasses
 import enum
 import functools
 from typing import TYPE_CHECKING, Any, Callable, Generic, Optional, TypedDict, TypeVar, cast
@@ -6,13 +5,15 @@ from typing import TYPE_CHECKING, Any, Callable, Generic, Optional, TypedDict, T
 import tmt.log
 import tmt.steps.provision
 import tmt.utils
-from tmt.plugins import PluginRegistry
-from tmt.utils import (
-    NormalizeKeysMixin,
+from tmt.container import (
     SerializableContainer,
     SpecBasedContainer,
+    container,
     field,
+    key_to_option,
     )
+from tmt.plugins import PluginRegistry
+from tmt.utils import NormalizeKeysMixin
 
 if TYPE_CHECKING:
     import tmt.base
@@ -102,7 +103,7 @@ class CheckResultInterpret(enum.Enum):
         return self.value
 
 
-@dataclasses.dataclass
+@container
 class Check(
         SpecBasedContainer[_RawCheck, _RawCheck],
         SerializableContainer,
@@ -145,7 +146,7 @@ class Check(
 
     def to_spec(self) -> _RawCheck:
         spec = cast(_RawCheck, {
-            tmt.utils.key_to_option(key): value
+            key_to_option(key): value
             for key, value in self.items()
             })
         spec["result"] = self.result.to_spec()
