+++ tmt/result.py
@@ -1,4 +1,3 @@
-import dataclasses
 import enum
 import re
 from typing import TYPE_CHECKING, Any, Callable, Optional, cast
@@ -11,7 +10,8 @@ import tmt.identifier
 import tmt.log
 import tmt.utils
 from tmt.checks import CheckEvent, CheckResultInterpret
-from tmt.utils import GeneralError, Path, SerializableContainer, field
+from tmt.container import SerializableContainer, container, field
+from tmt.utils import GeneralError, Path
 
 if TYPE_CHECKING:
     import tmt.base
@@ -125,7 +125,7 @@ ResultIds = dict[str, Optional[str]]
 RawResult = Any
 
 
-@dataclasses.dataclass
+@container
 class ResultGuestData(SerializableContainer):
     """
     Describes what tmt knows about a guest the result was produced on
@@ -163,7 +163,7 @@ def _unserialize_fmf_id(serialized: 'tmt.base._RawFmfId') -> 'tmt.base.FmfId':
     return FmfId.from_spec(serialized)
 
 
-@dataclasses.dataclass
+@container
 class BaseResult(SerializableContainer):
     """
     Describes what tmt knows about a result
@@ -217,7 +217,7 @@ class BaseResult(SerializableContainer):
         return ', '.join(self.note)
 
 
-@dataclasses.dataclass
+@container
 class CheckResult(BaseResult):
     """
     Describes what tmt knows about a single test check result
@@ -236,7 +236,7 @@ class CheckResult(BaseResult):
         return SubCheckResult.from_serialized(self.to_serialized())
 
 
-@dataclasses.dataclass
+@container
 class SubCheckResult(CheckResult):
     """
     Describes what tmt knows about a single subtest check result.
@@ -247,7 +247,7 @@ class SubCheckResult(CheckResult):
     """
 
 
-@dataclasses.dataclass
+@container
 class SubResult(BaseResult):
     """
     Describes what tmt knows about a single test subresult
@@ -261,14 +261,14 @@ class SubResult(BaseResult):
         )
 
 
-@dataclasses.dataclass
+@container
 class PhaseResult(BaseResult):
     """
     Describes what tmt knows about result of individual phases, e.g. prepare ansible
     """
 
 
-@dataclasses.dataclass
+@container
 class Result(BaseResult):
     """
     Describes what tmt knows about a single test result
