+++ tmt/steps/execute/__init__.py
@@ -20,6 +20,7 @@ import tmt.log
 import tmt.steps
 import tmt.utils
 from tmt.checks import CheckEvent
+from tmt.container import container, field
 from tmt.options import option
 from tmt.plugins import PluginRegistry
 from tmt.result import CheckResult, Result, ResultGuestData, ResultInterpret, ResultOutcome
@@ -31,7 +32,6 @@ from tmt.utils import (
     Path,
     ShellScript,
     Stopwatch,
-    field,
     format_duration,
     format_timestamp,
     )
@@ -244,7 +244,7 @@ SCRIPTS = (
     )
 
 
-@dataclasses.dataclass
+@container
 class ExecuteStepData(tmt.steps.WhereableStepData, tmt.steps.StepData):
     duration: str = field(
         # TODO: ugly circular dependency (see tmt.base.DEFAULT_TEST_DURATION_L2)
