+++ tmt/steps/report/__init__.py
@@ -1,4 +1,3 @@
-import dataclasses
 from typing import TYPE_CHECKING, Any, Optional, TypeVar, Union, cast
 
 import click
@@ -6,6 +5,7 @@ import click
 import tmt
 import tmt.plugins
 import tmt.steps
+from tmt.container import container
 from tmt.options import option
 from tmt.plugins import PluginRegistry
 from tmt.steps import Action
@@ -14,7 +14,7 @@ if TYPE_CHECKING:
     import tmt.cli
 
 
-@dataclasses.dataclass
+@container
 class ReportStepData(tmt.steps.StepData):
     pass
