+++ tmt/steps/report/reportportal.py
@@ -11,8 +11,9 @@ import tmt.hardware
 import tmt.log
 import tmt.steps.report
 import tmt.utils
+from tmt.container import container, field
 from tmt.result import ResultOutcome
-from tmt.utils import ActionType, catch_warnings_safe, field, format_timestamp, yaml_to_dict
+from tmt.utils import ActionType, catch_warnings_safe, format_timestamp, yaml_to_dict
 
 if TYPE_CHECKING:
     from tmt._compat.typing import TypeAlias
@@ -96,7 +97,7 @@ def _filter_log(log: str, settings: Optional[LogFilterSettings] = None) -> str:
     return log
 
 
-@dataclasses.dataclass
+@container
 class ReportReportPortalData(tmt.steps.report.ReportStepData):
 
     url: Optional[str] = field(
